llm/pyproject.toml
[project]
name = "astro-llm"
version = "0.1.0"
description = "Astro app LLM layer — Vedic astrology reading generation"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pyswisseph>=2.10",
    "timezonefinder>=6.0",
    "pytz>=2024.1",
    "geopy>=2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "rich>=13.0",
]
